requirements.txt
asgiref==3.3.1
dj-database-url==0.5.0
Django==3.1.7
django-crispy-forms==1.11.1
django-heroku==0.3.1
djangorestframework==3.12.2
environs==9.3.1
Markdown==3.3.3
psycopg2==2.8.6
pytz==2021.1
sqlparse==0.4.1
gunicorn==20.0.4
whitenoise==5.2.0